+++ Windows/WASAPIStream.cpp
@@ -127,12 +127,12 @@ public:
 	HRESULT STDMETHODCALLTYPE OnPropertyValueChanged(LPCWSTR pwstrDeviceId, const PROPERTYKEY key) override {
 		INFO_LOG(SCEAUDIO, "Changed audio device property "
 			"{%8.8x-%4.4x-%4.4x-%2.2x%2.2x-%2.2x%2.2x%2.2x%2.2x%2.2x%2.2x}#%d",
-			key.fmtid.Data1, key.fmtid.Data2, key.fmtid.Data3,
+			(uint32_t)key.fmtid.Data1, key.fmtid.Data2, key.fmtid.Data3,
 			key.fmtid.Data4[0], key.fmtid.Data4[1],
 			key.fmtid.Data4[2], key.fmtid.Data4[3],
 			key.fmtid.Data4[4], key.fmtid.Data4[5],
 			key.fmtid.Data4[6], key.fmtid.Data4[7],
-			key.pid);
+			(int)key.pid);
 		return S_OK;
 	}
